Implementation of the Time Statics Generator in order to create Hours for Reason of Visit

+++ app/src/main/java/generalinfo/powellmakerspace/org/makerspacelogin/utils/statistic/SqlTimeStatisticGenerator.java
@@ -2,6 +2,10 @@ package generalinfo.powellmakerspace.org.makerspacelogin.utils.statistic;
 
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.util.Log;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class SqlTimeStatisticGenerator implements StatisticGenerator {
 
@@ -18,11 +22,22 @@ public class SqlTimeStatisticGenerator implements StatisticGenerator {
     @Override
     public StatisticResult generateStatistic(){
 
-        long result;
-
+//        List<String> types = new ArrayList<String>();
+//        List<Integer> results = new ArrayList<Integer>();
+//
+//        try(Cursor c = db.rawQuery(sqlQuery, null)){
+//
+//
+//            while (c.moveToNext()){
+//                 types.add(c.getString(0));
+//                 results.add(c.getInt(1));
+//            }
+//
+//        }
+        int result;
         try(Cursor c = db.rawQuery(sqlQuery, null)){
             c.moveToNext();
-            result = c.getLong(0);
+            result = c.getInt(0);
         }
 
         return new TimeStatisticResult(fieldName, result);
